+++ examples/context_sharing.rs
@@ -80,7 +80,7 @@ impl InputContext for PlayerBox {
             .with_wasd()
             .with_modifier(Normalize)
             .with_modifier(DeltaLerp::default())
-            .with_modifier(Scalar::splat(DEFAULT_SPEED));
+            .with_modifier(Scale::splat(DEFAULT_SPEED));
 
         ctx.bind::<Rotate>().with(KeyCode::Space);