+++ examples/context_switch.rs
@@ -104,7 +104,7 @@ impl InputContext for OnFoot {
             .with_wasd()
             .with_modifier(Normalize)
             .with_modifier(DeltaLerp::default())
-            .with_modifier(Scalar::splat(DEFAULT_SPEED));
+            .with_modifier(Scale::splat(DEFAULT_SPEED));
         instance.bind::<Rotate>().with(KeyCode::Space);
         instance.bind::<EnterCar>().with(KeyCode::Enter);
 
@@ -135,7 +135,7 @@ impl InputContext for InCar {
             .with_wasd()
             .with_modifier(Normalize)
             .with_modifier(DeltaLerp::default())
-            .with_modifier(Scalar::splat(DEFAULT_SPEED + 200.0)); // Make car faster.
+            .with_modifier(Scale::splat(DEFAULT_SPEED + 200.0)); // Make car faster.
         ctx.bind::<ExitCar>().with(KeyCode::Enter);
 
         ctx